[rename] 패키지 명을 utils -> util 로 변경합니다.
