+++ src/main/java/team/waitingcatch/app/security/service/AccessDeniedHandlerImpl.java
@@ -10,7 +10,7 @@ import org.springframework.http.HttpStatus;
 import org.springframework.security.access.AccessDeniedException;
 import org.springframework.security.web.access.AccessDeniedHandler;
 
-import team.waitingcatch.app.security.utils.SecurityExceptionUtil;
+import team.waitingcatch.app.security.util.SecurityExceptionUtil;
 
 public class AccessDeniedHandlerImpl implements AccessDeniedHandler {
 	@Override
